+++ src/pages/authors.js
@@ -1,5 +1,5 @@
 import { gql } from "@apollo/client";
-import createApolloClient from "./apollo_client";
+import createApolloClient from "../_apollo_client";
 import Image from "next/image";
 import Link from "next/link";
